Examples: use cargo2nix from GitHub rather that from ../..
Rationale:

It's an extra mental load for users when they are shown one thing and
told to do another.

Use of parent directories in flake URLs is unreliable (that's why `get-flake`
exists) and should be avoided, especially in examples.

Referring to the parent directory creates a loop in flake locks. The
`flake.lock` files in examples affect the checksum of the parent flake,
so `nix flake lock` keeps changing checksums on every run.

With GitHub references, the locks only get out of date when a new commit
is merged to the branch used in the URL.